+++ Runtime/Data/UI/Popup/PopupButtonData.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace SadUtils
+{
+    public abstract class PopupButtonData
+    {
+        public abstract string Type { get; }
+        public abstract Action Callback { get; }
+    }
+}