+++ Runtime/Data/UI/Popup/PopupContentData.cs
@@ -1,0 +1,7 @@
+namespace SadUtils
+{
+    public abstract class PopupContentData
+    {
+        public abstract string Type { get; }
+    }
+}